api/requirements.txt
mailersend>=2.0.0
fastapi==0.124.0
uvicorn[standard]==0.35.0
psycopg2-binary==2.9.11
sqlalchemy==2.0.45
pydantic==2.12.5
python-jose==3.5.0
argon2-cffi==25.1.0
python-dotenv==1.0.0
boto3==1.35.0
python-multipart==0.0.6
google-auth==2.35.0
google-auth-oauthlib==1.2.1